Create blackberry project Part 1
Initial changes to support creating titanium projects for blackberry.  I needed to switch the args parsing to use argparse to make it easier to maintain and add new options when necessary

CHANGELOG
- replaced the command line parsing with an argparse based one
- added --android_sdk option to hopefully replace the optional last positional parameter in the future
- create the blackberry resources folder

TESTS
- Installed the modified project.py file to the mobile_sdk
- created a new project for android and mobileweb using Studio
- Verified that the projects still compile and run normally

Didn't test for iPhone as i don't have the setup, will ask someone with setup to give it a run.