Fix issue where the android_sdk long option value was not picked up.
Cause was that the optional and positional parameters for it were using the same destination and android_sdk would always be None.  As the positional one is only there to show in the usage messages, modified it's destination to notUsed.

TEST
Verified that the android-sdk param now gets picked up correctly, via debug log and also verified that the android files get generated for the project.